feat: Add Constant and IfExpression to grammar
The changes in this commit add support for constant declarations and if expressions in the grammar. The `Constant` variant is added to the `ToplevelStatement` and `Statement` enums, allowing for the declaration of constants. Additionally, the `IfExpression` variant is added to the `Expression` enum, enabling the use of if-else expressions in the language.

These changes are important to expand the functionality of the language and provide more flexibility in the types of constructs that can be expressed in the grammar.