feat: Add new language constructs to the grammar
This commit introduces several new language constructs to the grammar,
including `Declaration`, `StructDeclaration`, `Instance`, and
`MethodDeclaration`. These changes are part of the ongoing effort to
expand the capabilities of the language and provide a more
comprehensive set of features.

The key changes are:

- Added `Declaration` to represent variable declarations
- Added `StructDeclaration` to represent struct definitions
- Added `Instance` to represent instance declarations
- Added `MethodDeclaration` to represent methods within instances

These new constructs will enable the language to support a wider range
of use cases and provide a more robust type system and object-oriented
programming model.