feat(grammar): Add support for struct initialisation
This commit adds support for struct initialisation in the grammar.
It introduces a new `StructInitialisation` variant in the `Expression`
enum, as well as a `StructFieldInitialisation` struct to represent the
individual field initializations within a struct initializer.

The changes also include updates to the `ArrayType` struct to use a
single `&[` token instead of separate `&` and `[` tokens.

These changes are necessary to support the new struct initialisation
syntax in the language.